requirements.txt
h5py==3.7.0
lpips==0.1.4
matplotlib==3.6.2
medutils==0.1.21
Medutils.egg==info
numpy==1.23.5
pandas==1.5.2
pyexcel_ods==0.6.0
PyYAML==6.0.1
scipy==1.11.1
termcolor==2.3.0
timm==0.6.12
torch==1.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.64.1
